Update README.md
The model which is developed on the grid-based sampling system with 500x500m resolution as part of the research has six sub-models providing a calculation of the model variables. In order to operate the model, a few rasters and vector data are required and a geodatabase working area must be defined. Raster Data: (1) Landsat-8, Band4, (2) Landsat-8, Band5, (3) Landsat-8, Band10 Vector data: (1) Polygon boundary of urbanized area, (2) Sample grid polygon, (3) 2D Building shape polygons with the number of floors info, (4) 3D building multipatches.